[refactor] Printing methods have been moved to PrintQubit class
